Module Logic.Const

type t = const

val equal : t -> t -> bool

val hash : t -> int

val to_string : t -> string

val of_string : string -> t

val query : t

Special symbol, that will never occur in any user-defined clause or term. For strings, this may be the empty string "".