Module S.Subst

type t

type scope = int

type renaming

val empty : t

Empty subst

val bind : t -> T.t -> scope -> T.t -> scope -> t

Bind a variable,scope to a term,scope

val deref : t -> T.t -> scope -> T.t * scope

While the term is a variable bound in subst, follow its binding. Returns the final term and scope

val create_renaming : unit -> renaming

val reset_renaming : renaming -> unit

val rename : renaming:renaming -> T.t -> scope -> T.t

Rename the given variable into a variable that is unique within variables known to the given renaming

val eval : t -> renaming:renaming -> T.t -> scope -> T.t

Apply the substitution to the term. Free variables are renamed using renaming

val eval_lit : t -> renaming:renaming -> Lit.t -> scope -> Lit.t

val eval_lits : t -> renaming:renaming -> Lit.t list -> scope -> Lit.t list

val eval_clause : t -> renaming:renaming -> C.t -> scope -> C.t

val fmt : Stdlib.Format.formatter -> t -> unit