datalog index

Library datalog

The entry point of this library is the module: Datalog.

Library datalog.caml_interface

The entry point of this library is the module: Datalog_caml_interface.

Library datalog.top_down

The entry point of this library is the module: Datalog_top_down.

Library datalog.unix

The entry point of this library is the module: Datalog_unix.